refactor: split up config plugin from plugin mixin bootstrapper
