+++ plugin-mixin-bootstrapper/src/main/java/pw/kaboom/papermixins/pluginmixin/bootstrap/MixinExtrasConfigPlugin.java
@@ -1,37 +1,22 @@
-package pw.kaboom.papermixins.mixin;
+package pw.kaboom.papermixins.pluginmixin.bootstrap;
 
+import com.llamalad7.mixinextras.MixinExtrasBootstrap;
 import org.objectweb.asm.tree.ClassNode;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
 import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
-import pw.kaboom.papermixins.config.PaperMixinsConfig;
 
-import java.io.File;
 import java.util.List;
 import java.util.Set;
 
-public final class PaperMixinsMixinPlugin implements IMixinConfigPlugin {
-    private static final Logger LOGGER = LoggerFactory.getLogger("kaboom-paper-mixins");
-    private String mixinPackage;
-    private PaperMixinsConfig config;
-
+public final class MixinExtrasConfigPlugin implements IMixinConfigPlugin {
     @Override
     public void onLoad(final String mixinPackage) {
-        this.mixinPackage = mixinPackage + ".";
-        this.config = new PaperMixinsConfig(new File("config/kaboom-paper-mixins.properties"));
+        MixinExtrasBootstrap.init();
     }
 
     @Override
     public boolean shouldApplyMixin(final String targetClassName, final String mixinClassName) {
-        if (!mixinClassName.startsWith(this.mixinPackage)) {
-            LOGGER.warn("Expected mixin '{}' to start with package root '{}', treating as foreign and disabling!",
-                    mixinClassName, this.mixinPackage);
-            return false;
-        }
-
-        final String mixin = mixinClassName.substring(this.mixinPackage.length());
-        return this.config.canLoad(mixin);
+        return true;
     }
 
     @Override
